b3

+++ HMIS/src/java/cms/access/Access_User.java
@@ -106,7 +106,7 @@ public class Access_User {
             StringBuilder html = new StringBuilder();
             StringBuilder html3 = new StringBuilder();
 
-            DefaultTableModel dtm = db.Select(tableName, _id + "<>1");
+            DefaultTableModel dtm = db.Select(tableName, _id + ">5");// تا شماره  پنج کاربر های رزرو هستند
             List<Map<String, Object>> row = jjDatabase.separateRow(dtm);
             html.append(" <div class='card bd-primary mg-t-20'>"
                     + "    <div class='card-header bg-primary tx-white'>کاربران</div>"
@@ -153,7 +153,8 @@ public class Access_User {
             return "";
         }
     }
-           public static String changeStatus(HttpServletRequest request, HttpServletResponse response, jjDatabaseWeb db, String id, String newSatus) throws Exception {
+
+    public static String changeStatus(HttpServletRequest request, HttpServletResponse response, jjDatabaseWeb db, String id, String newSatus) throws Exception {
         try {
             String errorMessageId = jjValidation.isDigitMessageFa(id, "کد");
             if (!errorMessageId.equals("")) {
@@ -196,7 +197,8 @@ public class Access_User {
                 html.append(Js.setHtml("#User_button", ""));
             }
 
-            List<Map<String, Object>> row = jjDatabase.separateRow(db.Select(tableName));
+            List<Map<String, Object>> row = jjDatabase.separateRow(db.Select(tableName, _id + "=0"));
+            //برای نشان دادن لوگوی اختصاصی بیمارستان در قسمت اضافه کردن کاربان
             StringBuilder script2 = new StringBuilder();
             if (row.get(0).get(Access_User._attachPicPersonal).equals("")) {
                 script2.append(Js.setAttr("#PicPreviewPersonal", "src", "img/preview.jpg"));
@@ -479,8 +481,13 @@ public class Access_User {
                 }
                 Server.outPrinter(request, response, Js.dialog(errorMessageId));
                 return "";
+            }            
+            if (id.equals("0") || id.equals("1") || id.equals("2") || id.equals("3") || id.equals("4") || id.equals("1")) {
+                String errorMessage = "شما اجازه مشاهده اطلاعات این شخص را ندارید";
+                Server.outPrinter(request, response, Js.dialog(errorMessage) + Js.jjUser.showTbl());
+                return "";
             }
-
+           //@ToDo کاربرانی که در قسمت های مختلف سیستم تراکنش داشته اند را نابید بتوانیم خذف کنیم
             if (!db.delete(tableName, _id + "=" + id)) {
                 String errorMessage = "عملیات حذف به درستی صورت نگرفت";
                 if (jjTools.isLangEn(request)) {
@@ -509,7 +516,7 @@ public class Access_User {
      *
      * @param id
      */
-     public static String select(HttpServletRequest request, HttpServletResponse response, jjDatabaseWeb db, boolean needString) throws Exception {
+    public static String select(HttpServletRequest request, HttpServletResponse response, jjDatabaseWeb db, boolean needString) throws Exception {
         try {
             String id = jjTools.getParameter(request, _id);
             String errorMessageId = jjValidation.isDigitMessageFa(id, "کد");
@@ -521,7 +528,7 @@ public class Access_User {
                 return "";
             }
 
-            if (id.equals("1")) {
+            if (id.equals("0") || id.equals("1") || id.equals("2") || id.equals("3") || id.equals("4") || id.equals("1")) {
                 String errorMessage = "شما اجازه مشاهده اطلاعات این شخص را ندارید";
                 Server.outPrinter(request, response, Js.dialog(errorMessage) + Js.jjUser.showTbl());
                 return "";
@@ -546,7 +553,7 @@ public class Access_User {
             html.append(Js.setVal("#user_nameUser", row.get(0).get(_name)));
             html.append(Js.setVal("#user_AccountInformationUser", row.get(0).get(_AccountInformation)));
 
-            html.append(Js.setVal("#user_birthdateUser", row.get(0).get(_birthdate)));
+            html.append(Js.setVal("#user_birthdateUser", jjCalendar_IR.getViewFormat(row.get(0).get(_birthdate))));
             html.append(Js.setVal("#user_gradeUser", row.get(0).get(_grade)));
             html.append(Js.setVal("#user_jensiatUser", row.get(0).get(_jensiat)));
             html.append(Js.setVal("#user_codeMeliUser", row.get(0).get(_codeMeli)));
@@ -560,10 +567,6 @@ public class Access_User {
             html.append(Js.setAttr("#DownloadPicPersonnelCard", "href", "upload/" + row.get(0).get(_attachPicPersonnelCard)));
             /////برای دانلود عکس ها نوشته شده
             html.append(Js.setAttr("#DownloadPicSignature", "href", "upload/" + row.get(0).get(_attachPicSignature)));
-
-//            html.append(Js.setHtml("#user_pic1", row.get(0).get(_attachAxPersonal)));
-//            html.append(Js.setHtml("#user_pic3", row.get(0).get(_attachAxPersonnelCard)));
-//            html.append(Js.setHtml("#user_pic2", row.get(0).get(_attachAxSignature)));
             String script1 = "";
             StringBuilder html3 = new StringBuilder();
             StringBuilder html4 = new StringBuilder();
@@ -573,24 +576,54 @@ public class Access_User {
             String attachFiles = row.get(0).get(_attachFile).toString();
 
             String[] attachFilesArray = attachFiles.split("#A#");
-
-//      if (row.get(0).get(Access_User._attachFile).equals("")) {
-//      html4.append("$('#inputAfterSelect').hide()");}
-//      else{
-//           html4.append("$('#inputAfterSelect').show()");
-            for (int l = 0; l < attachFilesArray.length; l++) {
+//
+////            for (int l = 0; l < attachFilesArray.length; l++) {
+////                List<Map<String, Object>> fileRow = jjDatabase.separateRow(db.Select(UploadServlet.tableName, UploadServlet._file_name + "='" + attachFilesArray[l] + "'"));
+////                if (!fileRow.isEmpty()) {
+////                    String idUpload = fileRow.get(0).get(UploadServlet._id).toString();
+//////                    html3.append("<div >" + "<input class='col-xs-12' disabled='disabled'  value='" + attachFilesArray[l] + "'/>"  + "</div>");
+////                    html3.append( "<input style='text-align: center;' class='col-lg-11' disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<div  onclick='cmsUser.m_remove(" + idUpload + "," + id + ")'>" + "<img  src='imgfeyz/delet.png' style='width:2%' />" + "</div>" );
+////
+//////                    html3.append("<div class='col-xs-12'>" + "<input  disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<div  onclick='"+Js.modal("hjgjhgkjgkjgk", id)+"'>" + "<img  src='imgfeyz/delet.png' style='width:2%' />" + "</div>" + "</div>");
+////         
+////                }
+////            }
+//
+//            script1 = Js.setHtml("#inputAfterSelect", html3);
+            
+            //////////////
+            for (int l = 0; l < attachFilesArray.length ; l++) {
                 List<Map<String, Object>> fileRow = jjDatabase.separateRow(db.Select(UploadServlet.tableName, UploadServlet._file_name + "='" + attachFilesArray[l] + "'"));
+                
                 if (!fileRow.isEmpty()) {
                     String idUpload = fileRow.get(0).get(UploadServlet._id).toString();
-//                    html3.append("<div >" + "<input class='col-xs-12' disabled='disabled'  value='" + attachFilesArray[l] + "'/>"  + "</div>");
-                    html3.append( "<input style='text-align: center;' class='col-lg-11' disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<div  onclick='cmsUser.m_remove(" + idUpload + "," + id + ")'>" + "<img  src='imgfeyz/delet.png' style='width:2%' />" + "</div>" );
+                    String titleUpload = fileRow.get(0).get(UploadServlet._title).toString();
+                    String extension2 = attachFilesArray[l].substring(attachFilesArray[l].lastIndexOf(".") + 1, attachFilesArray[l].length());
+                    if (extension2.toLowerCase().equals("jpg")
+                            || extension2.toLowerCase().equals("png")
+                            || extension2.toLowerCase().equals("gif")
+                            || extension2.toLowerCase().equals("svg")) {
+                        if (titleUpload.equals("")) {
+                            html2.append("<img class='col-xs-12' style='width:10%;float:right' src='upload/" + attachFilesArray[l] + "'/><input style='text-align: center;' class='col-lg-12'  disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<button class='col-lg-1 form-control'  style='background-color: #e16262;color: white;float:left' onclick='cmsUser.m_remove(" + idUpload + "," + id + ")'>" + "حذف" + "</button><a  class='col-lg-1' style='background-color: green;color: white;float:left;text-align: center;padding-top: 5px;padding-bottom: 9px;margin-top: 1px;' href='upload/" + attachFilesArray[l] + "' >دانلود</a>");
+                        } else {
 
-//                    html3.append("<div class='col-xs-12'>" + "<input  disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<div  onclick='"+Js.modal("hjgjhgkjgkjgk", id)+"'>" + "<img  src='imgfeyz/delet.png' style='width:2%' />" + "</div>" + "</div>");
-         
+                            html2.append("<img class='col-xs-12' style='width:10%;float:right' src='upload/" + attachFilesArray[l] + "'/><input class='col-lg-12 form-control'  style='text-align: center' disabled='disabled'  value='" + titleUpload+ "'/>" + "<input  style='text-align: center;' class='col-lg-12'  disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<button  class='col-lg-1' style='background-color: #e16262;color: white;float:left' onclick='cmsUser.m_remove(" + idUpload + "," + id + ")'>حذف" + "</button><a  class='col-lg-1' style='background-color: green;color: white;float:left;text-align: center;padding-top: 2px;padding-bottom: 1px;margin-top: 1px;' href='upload/" + attachFilesArray[l] + "' >دانلود</a>");
+                        }
+                    } else {
+                        html2.append("<input class='col-lg-12 form-control'  style='text-align: center' disabled='disabled'  value='" + titleUpload + "'/>" + "<input  style='text-align: center;' class='col-lg-12'  disabled='disabled'  value='" + attachFilesArray[l] + "'/>" + "<button  class='col-lg-1' style='background-color: #e16262;color: white;float:left' onclick='cmsUser.m_remove(" + idUpload + "," + id + ")'>حذف" + "</button><a  class='col-lg-1' style='background-color: green;color: white;float:left;text-align: center;padding-top: 2px;padding-bottom: 1px;margin-top: 1px;' href='upload/" + attachFilesArray[l] + "' >دانلود</a>");
+                    }
                 }
             }
 
-            script1 = Js.setHtml("#inputAfterSelect", html3);
+            script1 += Js.setHtml(".inputAfterSelect", html2);
+            //////////////
+            
+            
+            
+            
+            
+            
+            
 
             if (row.get(0).get(Access_User._attachPicPersonal).equals("")) {
                 script.append(Js.setAttr("#PicPreviewPersonal", "src", "img/preview.jpg"));
@@ -613,7 +646,7 @@ public class Access_User {
             }
 
             html.append(Js.setVal("#user_addressUser", row.get(0).get(_address)));
-            html.append(Js.setValDate("#user_birthdateUserUser", row.get(0).get(_birthdate)));
+            html.append(Js.setValDate("#user_birthdateUserUser", jjCalendar_IR.getViewFormat(row.get(0).get(_birthdate))));
 ///////////////////////////
             /////این تابع برای نمایش فایل های اپلود شده توسط فردی که واردشده نوشته شده است
             /////شیران1
@@ -734,7 +767,7 @@ public class Access_User {
             System.out.println("____________________________________");
 
             db.update(tableName, map, _id + "=" + idUser);
-            changeStatus(request, response, db, idUpload,status_deleted+" "+jjTools.getSessionAttribute(request, "#USER_NAME")+" "+jjTools.getSessionAttribute(request, "#USER_FAMILY"));
+            changeStatus(request, response, db, idUpload, status_deleted + " " + jjTools.getSessionAttribute(request, "#USER_NAME") + " " + jjTools.getSessionAttribute(request, "#USER_FAMILY"));
 //            if (!db.delete(UploadServlet.tableName, UploadServlet._id + "=" + idUpload)) {
 //                String errorMessage = "عملیات حذف به درستی صورت نگرفت";
 //                if (jjTools.isLangEn(request)) {
@@ -1524,7 +1557,8 @@ public class Access_User {
             return "";
         }
     }
-        /**
+
+    /**
      * این متد کاربران فعال را بصورت آپشن برای قرار گرفتن در سلکت بر می گرداند
      *
      * @param request panel درون ریکوئست اگر با نقطه شروع نشود آی دی در نظر می
@@ -1532,22 +1566,23 @@ public class Access_User {
      * @param response
      * @param db
      * @param needString
-     * @return بصورت کد جی کوئری و یک سری آپشن برای قرار گرفتن در سلکتی که در پنل معرفی شده
+     * @return بصورت کد جی کوئری و یک سری آپشن برای قرار گرفتن در سلکتی که در
+     * پنل معرفی شده
      * @throws Exception
      */
     public static String getSelectOption(HttpServletRequest request, HttpServletResponse response, jjDatabaseWeb db, boolean needString) throws Exception {
         StringBuilder optionHtml = new StringBuilder();
         try {
-            List<Map<String, Object>> rowAllActiveRols = jjDatabase.separateRow(db.Select(tableName, _id + "," + _name+ "," +_family , "id>0 AND "+_isActive+"=1", _family));// بر اساس حروف الفبا مرتب باشد بهتر است
-                optionHtml.append("<option  value='ALL'>تمام کاربران ثبت شده</option>");
+            List<Map<String, Object>> rowAllActiveRols = jjDatabase.separateRow(db.Select(tableName, _id + "," + _name + "," + _family, "id>0 AND " + _isActive + "=1", _family));// بر اساس حروف الفبا مرتب باشد بهتر است
+            optionHtml.append("<option  value='ALL'>تمام کاربران ثبت شده</option>");
             for (int i = 0; i < rowAllActiveRols.size(); i++) {
-                optionHtml.append("<option  value='").append(rowAllActiveRols.get(i).get(_id)).append("'>").append(rowAllActiveRols.get(i).get(_family)+"-").append(rowAllActiveRols.get(i).get(_name)).append("</option>");
+                optionHtml.append("<option  value='").append(rowAllActiveRols.get(i).get(_id)).append("'>").append(rowAllActiveRols.get(i).get(_family) + "-").append(rowAllActiveRols.get(i).get(_name)).append("</option>");
             }
             String panel = jjTools.getParameter(request, "panel");
             if (panel.isEmpty()) {
                 panel = ".usersSelectOption";
             }
-            Server.outPrinter(request, response,  Js.setHtml(panel, optionHtml));
+            Server.outPrinter(request, response, Js.setHtml(panel, optionHtml));
             return "";
         } catch (Exception e) {
             Server.outPrinter(request, response, Server.ErrorHandler(e));
